# Conflicts:
#	Unreal/.beamable/content-snapshots/LastPublished-global.json
#	Unreal/BeamableServices/services/StellarFederation/StellarFederation.cs
#	Unreal/Content/Game/Blueprints/Beam/WBP_MainMenu.uasset
#	Unreal/Content/Game/Blueprints/Characters/Redhood/ABP_Redhood.uasset
#	Unreal/Content/Game/Blueprints/Characters/Redhood/BP_RedHood.uasset
#	Unreal/Plugins/Unreal2dDungeonMicroserviceClients/Source/Unreal2dDungeonMicroserviceClients/Private/AutoGen/SubSystems/BeamStellarFederationApi.cpp
#	Unreal/Plugins/Unreal2dDungeonMicroserviceClients/Source/Unreal2dDungeonMicroserviceClients/Public/AutoGen/SubSystems/BeamStellarFederationApi.h
#	Unreal/Plugins/Unreal2dDungeonMicroserviceClients/Unreal2dDungeonMicroserviceClients_GenerationPass.json

+++ Unity/BeamableServices/services/StellarFederation/Features/Scheduler/Storage/Modles/Block.cs
@@ -6,5 +6,5 @@ public class Block
 {
     [BsonElement("_id")]
     public string Network { get; set; } = null!;
-    public long BlockNumber { get; set; }
+    public uint BlockNumber { get; set; }
 }
